+++ clients/admin-ui/src/features/custom-fields/CustomFieldsTable.tsx
@@ -1,26 +1,43 @@
-import Restrict, { useHasPermission } from "common/Restrict";
+/* eslint-disable react/no-unstable-nested-components */
 import {
-  FidesTable,
-  FidesTableFooter,
-  TitleCell,
-  WrappedCell,
-} from "common/table";
-import EmptyTableState from "common/table/EmptyTableState";
+  ColumnDef,
+  createColumnHelper,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getSortedRowModel,
+  useReactTable,
+} from "@tanstack/react-table";
 import {
   Box,
+  BoxProps,
   Button,
-  Flex,
-  Spinner,
+  HStack,
   Text,
   useDisclosure,
   useToast,
+  VStack,
 } from "fidesui";
-import React, { useMemo, useState } from "react";
-import { Column, Hooks } from "react-table";
+import { useMemo, useState } from "react";
 
 import { useAppSelector } from "~/app/hooks";
 import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
+import Restrict, { useHasPermission } from "~/features/common/Restrict";
+import {
+  DefaultHeaderCell,
+  FidesTableV2,
+  GlobalFilterV2,
+  TableActionBar,
+  TableSkeletonLoader,
+} from "~/features/common/table/v2";
 import { errorToastParams, successToastParams } from "~/features/common/toast";
+import {
+  EnableCustomFieldCell,
+  FieldTypeCell,
+  ResourceTypeCell,
+  ValueTextCell,
+} from "~/features/custom-fields/cells";
+import { CustomFieldActions } from "~/features/custom-fields/CustomFieldActions";
+import { CustomFieldModal } from "~/features/custom-fields/CustomFieldModal";
 import {
   selectAllCustomFieldDefinitions,
   useDeleteCustomFieldDefinitionMutation,
@@ -28,25 +45,46 @@ import {
 } from "~/features/plus/plus.slice";
 import { CustomFieldDefinitionWithId, ScopeRegistryEnum } from "~/types/api";
 
-import {
-  EnableCustomFieldCell,
-  FieldTypeCell,
-  MoreActionsCell,
-  ResourceTypeCell,
-} from "./cells";
-import { CustomFieldModal } from "./CustomFieldModal";
+const ADD_CUSTOM_FIELD_BTN_LABEL = "Add a custom field";
 
-export const CustomFieldsTable = () => {
+const EmptyTableNotice = () => (
+  <VStack
+    mt={6}
+    p={10}
+    spacing={4}
+    borderRadius="base"
+    maxW="70%"
+    data-testid="no-results-notice"
+    alignSelf="center"
+    margin="auto"
+    textAlign="center"
+  >
+    <VStack>
+      <Text fontSize="md" fontWeight="600">
+        It looks like it’s your first time here!
+      </Text>
+      <Text fontSize="sm">
+        You haven’t created any custom fields yet.
+        <br />
+        To create a custom field, click on the{" "}
+        <strong>&quot;{ADD_CUSTOM_FIELD_BTN_LABEL}&quot;</strong> button
+      </Text>
+    </VStack>
+  </VStack>
+);
+
+const columnHelper = createColumnHelper<CustomFieldDefinitionWithId>();
+
+export const CustomFieldsTable = ({ ...rest }: BoxProps): JSX.Element => {
   const toast = useToast();
   const { isLoading } = useGetAllCustomFieldDefinitionsQuery();
   const customFields = useAppSelector(selectAllCustomFieldDefinitions);
   const [deleteCustomFieldDefinitionMutationTrigger] =
     useDeleteCustomFieldDefinitionMutation();
   const { isOpen, onClose, onOpen } = useDisclosure();
-
-  const [activeCustomField, setActiveCustomField] = useState<
-    CustomFieldDefinitionWithId | undefined
-  >(undefined);
+  const [activeCustomField, setActiveCustomField] =
+    useState<CustomFieldDefinitionWithId>();
+  const [globalFilter, setGlobalFilter] = useState<string>();
 
   // Permissions
   const userCanUpdate = useHasPermission([
@@ -76,148 +114,130 @@ export const CustomFieldsTable = () => {
     }
   };
 
-  const tableHook = (hooks: Hooks<CustomFieldDefinitionWithId>) => {
-    if (!userCanUpdate && !userCanDelete) {
-      return;
-    }
-    hooks.visibleColumns.push((visibleColumns) => [
-      ...visibleColumns,
-      {
-        id: "more-actions",
-        Header: () => null,
-        Cell: MoreActionsCell,
-        width: "50px",
-        onEdit: handleRowClick,
-        onDelete: handleDelete,
-      },
-    ]);
-  };
+  const columns: ColumnDef<CustomFieldDefinitionWithId, any>[] = useMemo(
+    () =>
+      [
+        columnHelper.accessor((row) => row.name, {
+          id: "name",
+          cell: ValueTextCell,
+          header: (props) => <DefaultHeaderCell value="Label" {...props} />,
+        }),
+        columnHelper.accessor((row) => row.description, {
+          id: "description",
+          cell: ValueTextCell,
+          header: (props) => (
+            <DefaultHeaderCell value="Description" {...props} />
+          ),
+        }),
+        columnHelper.accessor((row) => row.field_type, {
+          id: "field_type",
+          cell: FieldTypeCell,
+          header: (props) => (
+            <DefaultHeaderCell value="Field Type" {...props} />
+          ),
+        }),
+        columnHelper.accessor((row) => row.resource_type, {
+          id: "resource_type",
+          cell: ResourceTypeCell,
+          header: (props) => <DefaultHeaderCell value="Locations" {...props} />,
+        }),
+        userCanUpdate &&
+          columnHelper.accessor((row) => row.active, {
+            id: "enable",
+            cell: EnableCustomFieldCell,
+            header: (props) => <DefaultHeaderCell value="Enable" {...props} />,
+            meta: { disableRowClick: true },
+          }),
+        (userCanUpdate || userCanDelete) &&
+          columnHelper.display({
+            id: "actions",
+            header: "Actions",
+            cell: ({ row }) => (
+              <CustomFieldActions
+                customField={row.original}
+                onEdit={handleRowClick}
+                onDelete={handleDelete}
+              />
+            ),
+            meta: { disableRowClick: true },
+          }),
+      ].filter(Boolean) as ColumnDef<CustomFieldDefinitionWithId, any>[],
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [userCanDelete, userCanUpdate]
+  );
+
+  const tableInstance = useReactTable<CustomFieldDefinitionWithId>({
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    onGlobalFilterChange: setGlobalFilter,
+    globalFilterFn: "includesString",
+    columns,
+    data: customFields,
+    state: {
+      globalFilter,
+    },
+  });
 
   const handleCloseModal = () => {
     setActiveCustomField(undefined);
     onClose();
   };
 
-  const columns: Column<CustomFieldDefinitionWithId>[] = useMemo(
-    () => [
-      {
-        Header: "Label",
-        accessor: "name",
-        Cell: TitleCell,
-      },
-      {
-        Header: "Description",
-        accessor: "description",
-        Cell: WrappedCell,
-      },
-      { Header: "Field Type", accessor: "field_type", Cell: FieldTypeCell },
-      {
-        Header: "Locations",
-        accessor: "resource_type",
-        Cell: ResourceTypeCell,
-      },
-      {
-        Header: "Enable",
-        accessor: (row) => !row.active,
-        disabled: !userCanUpdate,
-        Cell: EnableCustomFieldCell,
-      },
-    ],
-    [userCanUpdate]
-  );
-
-  if (isLoading) {
-    return (
-      <Flex height="100%" justifyContent="center" alignItems="center">
-        <Spinner />
-      </Flex>
-    );
-  }
-
-  const addCustomFieldButton = (
+  const AddCustomFieldButton = () => (
     <Restrict scopes={[ScopeRegistryEnum.CUSTOM_FIELD_DEFINITION_CREATE]}>
       <Button
-        size={customFields.length > 0 ? "xs" : "sm"}
-        colorScheme={customFields.length > 0 ? "primary" : undefined}
-        variant={customFields.length === 0 ? "outline" : undefined}
-        fontWeight={customFields.length === 0 ? "semibold" : undefined}
-        minWidth={customFields.length === 0 ? "auto" : undefined}
+        size="xs"
+        colorScheme="primary"
         data-testid="add-custom-field-btn"
         onClick={onOpen}
       >
-        Add a custom field +
+        {ADD_CUSTOM_FIELD_BTN_LABEL}
       </Button>
     </Restrict>
   );
 
-  if (customFields.length === 0) {
-    return (
-      <Box maxWidth="720px">
-        <Text fontSize="sm" lineHeight={5} fontWeight="medium" marginBottom={6}>
-          Custom fields provide organizations with the capability to capture
-          metrics that are unique to their specific needs, allowing them to
-          create customized reports. These fields can be added to either systems
-          or elements within a taxonomy, and once added, they become reportable
-          fields that are visible on the data map.
-        </Text>
-        <EmptyTableState
-          title="It looks like it’s your first time here!"
-          button={addCustomFieldButton}
-          description={
-            <>
-              You haven’t created any custom fields yet. To create a custom
-              field, click on the{" "}
-              <strong>&quot;Add a custom field&quot;</strong> button
-            </>
-          }
-        />
-
-        {isOpen ? (
-          <CustomFieldModal
-            customField={activeCustomField}
-            isOpen={isOpen}
-            onClose={handleCloseModal}
-            isLoading={false}
-          />
-        ) : null}
-      </Box>
-    );
-  }
-
   return (
-    <>
-      <Box maxWidth="600px">
-        <Text marginBottom={10} fontSize="sm">
+    <Box {...rest}>
+      <Box maxWidth={600}>
+        <Text mb={10} fontSize="sm">
           {customFields.length > 0
             ? "Custom fields enable you to capture metrics specific to your organization and have those metrics appear in the data map and in reports."
             : "Custom fields provide organizations with the capability to capture metrics that are unique to their specific needs, allowing them to create customized reports. These fields can be added to either systems or elements within a taxonomy, and once added, they become reportable fields that are visible on the data map."}
         </Text>
       </Box>
-      <Box padding={2} data-testid="custom-fields-page">
-        <FidesTable<CustomFieldDefinitionWithId>
-          columns={columns}
-          data={customFields}
-          showSearchBar
-          searchBarPlaceHolder="Search for a custom field"
-          onRowClick={userCanUpdate ? handleRowClick : undefined}
-          customHooks={[tableHook]}
-          footer={
-            // +1 for total columns because our hook adds a column for the
-            // more actions menu
-            <FidesTableFooter totalColumns={columns.length + 1}>
-              {addCustomFieldButton}
-            </FidesTableFooter>
-          }
-        />
-        {isOpen ? (
+      <Box data-testid="custom-fields-page">
+        <TableActionBar>
+          <GlobalFilterV2
+            globalFilter={globalFilter}
+            setGlobalFilter={setGlobalFilter}
+            placeholder="Search for a custom field"
+          />
+          <HStack alignItems="center" spacing={4}>
+            <AddCustomFieldButton />
+          </HStack>
+        </TableActionBar>
+        {isLoading ? (
+          <Box p={2} borderWidth={1}>
+            <TableSkeletonLoader rowHeight={26} numRows={10} />
+          </Box>
+        ) : (
+          <FidesTableV2
+            tableInstance={tableInstance}
+            onRowClick={userCanUpdate ? handleRowClick : undefined}
+            emptyTableNotice={<EmptyTableNotice />}
+            enableSorting
+          />
+        )}
+        {isOpen && (
           <CustomFieldModal
             customField={activeCustomField}
             isOpen={isOpen}
             onClose={handleCloseModal}
             isLoading={false}
           />
-        ) : null}
+        )}
       </Box>
-    </>
+    </Box>
   );
 };